Удалил неиспользуемые функции
