+++ js/helper.js
@@ -18,17 +18,4 @@ function enableForm(formData, classData) {
   });
 }
 
-// Устранение дребезга
-function debounce(callback, delay = 500) {
-  let timeoutId;
-
-  return (...args) => {
-    clearTimeout(timeoutId);
-    timeoutId = setTimeout(() => {
-      timeoutId = null;
-      callback(...args);
-    }, delay);
-  }
-}
-
-export { disableForm, enableForm, debounce };
+export { disableForm, enableForm };
